Ability to run gbs under different user/group
This change adds new configuration file options to set the user/group
under which gbs is run in the service. The service itself is run under
whatever user/group the source service server itself is configured to
use. The service now spawns a separate thread for gbs and sets the
uid/gid there before calling gbs - this is accomplished by utilizing
functionality from the obs-service-git-buildpackage-utils module.

Change-Id: I0c8a9c4d2349cd0ca629f816cb9011f26a42b3bb
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>