Implement 'error-pkg' parameter
Instead of causing a service error generate a special "error package"
that shows the service error log but fails to build. The parameter value
is a comma-separated list of integers, representing the exit codes for
which the special error package is created.

This is a special hack to prevent the +creation of broken packages in
case of service failures.

Change-Id: I20b1e810b498a120245fe94276c180b458ab0b9a
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>